refactor: move transformStyle to compiler-dom
