+++ packages/compiler-dom/src/index.ts
@@ -1,6 +1,7 @@
 import { baseCompile, CompilerOptions, CodegenResult } from '@vue/compiler-core'
 import { parserOptionsMinimal } from './parserOptionsMinimal'
 import { parserOptionsStandard } from './parserOptionsStandard'
+import { transformStyle } from './transforms/transformStyle'
 
 export function compile(
   template: string,
@@ -9,6 +10,7 @@ export function compile(
   return baseCompile(template, {
     ...options,
     ...(__BROWSER__ ? parserOptionsMinimal : parserOptionsStandard),
+    nodeTransforms: [transformStyle, ...(options.nodeTransforms || [])],
     directiveTransforms: {
       // TODO include DOM-specific directiveTransforms
       ...(options.directiveTransforms || {})
