+++ packages/compiler-dom/src/transforms/vText.ts
